[UPDATE] v0.4.0 - ServerModal operation and global server config
